server/requirements.txt
asgiref==3.3.1
certifi==2020.6.20
Django==3.1.3
djangorestframework==3.12.1
gunicorn==20.0.4
pip==20.2.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
setuptools==50.3.0
sqlparse==0.4.1
wheel==0.35.1
whitenoise==5.2.0
